requirements.txt
aniso8601==9.0.1
click==8.0.4
Flask==2.0.3
Flask-RESTful==0.3.9
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
mysql-connector-python==8.0.28
protobuf==3.19.4
pytz==2021.3
six==1.16.0
Werkzeug==2.0.3
